Create a Whistleblowers policy for Privacytools.
in order to develope a proper policy for safe whistleblowing, i wrote this early draft from which we can move further.

I propose as written above two create two roles, and internal, and an external compliance officer. By default, complaints and reports should be send to the internal complaince officer, which should be dealing with the issue. However if the person with the complaint or report feels uncomfortable with talking to the internal officer, they may choose to report to the external compliance officer.  An example for the need for this is the situation where the complaint is about the internal complaince officer themself.

I would think that as community manager, it would be fairly appropriate if I became the internal complaince officer, but we still have to get an external complaince officer, which should be someone from outside of the PTIO team, but is still a trusted, but independent, person or organization.

Im looking forward to develope this into a mature policy and guideline, and role this out :)